tests: make the object-size fixture data-model portable
`unsigned long` is 32-bit under LLP64, so (usize)-1 no longer matched
the 64-bit (size_t)-1 the builtin yields on the Windows targets.

+++ tests/fixtures/c/builtin_object_size.c
@@ -2,7 +2,7 @@
  * declared array folds to its byte count; an unknown pointer yields
  * (size_t)-1 for the maximum forms (types 0 and 1) and 0 for the
  * minimum forms (types 2 and 3). The pointer operand is unevaluated. */
-typedef unsigned long usize;
+typedef unsigned long long usize;
 
 static char buf[16];
